added CNN .h5 file
This is not even the final version of the AI model yet. I understand that it has been a really long time since I was asked to document it, but ever since then, I've been experimenting with different ways to train the model in order to correct orientation. Progresses are made, and I took quite some notes. The discoveries will probably be mentioned in the paper (or, maybe these discoveries are too insignificant to AI researcher,s so I will mention them somewhere else).